Tag filter support for Mocha. Move Nightwatch tag filter to a better place.
