+++ src/interop/vanilla-mocha/tag_filter.js
@@ -1,0 +1,1 @@
+module.exports = require("../lib/mocha_tag_filter");
